Always offer Save Workspace as Layout even when the plus menu is otherwise empty
Review finding: the section model only appended the management section when
earlier sections or layout-management metadata existed, so a config with no
create/cloud/layout/template rows dropped the Save affordance entirely and
makeNewWorkspaceContextMenu returned nil. Append the management section
unconditionally, matching the pre-model menu. Regression test moved model
tests into NewWorkspaceMenuModelTests.swift to stay under the file length
budget.
