+++ cmuxTests/CmuxConfigNewWorkspaceMenuTests.swift
@@ -336,82 +336,6 @@ struct CmuxConfigNewWorkspaceMenuTests {
     }
 
     @MainActor
-    @Test func newWorkspaceMenuModelClassifiesAndOrdersSections() throws {
-        let (store, root) = try loadStore(globalJSON: """
-        {
-          "actions": {
-            "terminal-command": {
-              "type": "command",
-              "title": "Terminal Command",
-              "command": "echo hi"
-            },
-            "review-layout": {
-              "type": "workspace",
-              "title": "Review Layout",
-              "workspace": { "name": "Review" }
-            }
-          },
-          "ui": {
-            "newWorkspace": {
-              "menuSectionOrder": "customFirst",
-              "contextMenu": [
-                "cmux.newWorkspace",
-                { "type": "separator" },
-                "review-layout",
-                "terminal-command"
-              ],
-              "action": "terminal-command"
-            }
-          }
-        }
-        """)
-        defer { try? FileManager.default.removeItem(at: root) }
-
-        let agent = CmuxResolvedConfigAction.builtIn(.newAgentChat)
-        let model = NewWorkspaceMenuModel.build(
-            newWorkspaceContextMenuItems: store.newWorkspaceContextMenuItems,
-            agentChatAction: agent,
-            cloudSectionEnabled: true,
-            templateNames: ["Template A"],
-            loadedActions: store.loadedActions,
-            newWorkspaceActionID: store.newWorkspaceActionID,
-            deletable: { $0.id == "review-layout" },
-            sectionOrder: store.newWorkspaceMenuSectionOrder
-        )
-
-        guard case .create(let createRows) = model.sections[0],
-              case .cloud = model.sections[1],
-              case .layouts(let layoutRows) = model.sections[2],
-              case .templates(let templates) = model.sections[3],
-              case .management(let management) = model.sections[4] else {
-            Issue.record("Unexpected model sections: \(model.sections)")
-            return
-        }
-        let createIDs = createRows.compactMap { row -> String? in
-            guard case .action(let action, _, _) = row else { return nil }
-            return action.action.id
-        }
-        #expect(createIDs == [
-            CmuxSurfaceTabBarBuiltInAction.newWorkspace.configID,
-            "terminal-command",
-            CmuxSurfaceTabBarBuiltInAction.newAgentChat.configID,
-        ])
-        #expect(createRows.contains(.separator))
-        #expect(layoutRows.map { $0.menuAction.action.id } == ["review-layout"])
-        #expect(layoutRows.first?.deletable == true)
-        #expect(templates == ["Template A"])
-        #expect(management.deletableActions.map(\.id) == ["review-layout"])
-        if case .action(_, _, let isDefault)? = createRows.first(where: { row in
-            guard case .action(let action, _, _) = row else { return false }
-            return action.action.id == "terminal-command"
-        }) {
-            #expect(isDefault)
-        } else {
-            Issue.record("Expected hand-edited default command in create rows")
-        }
-    }
-
-    @MainActor
     @Test func storeRespectsNewWorkspaceMenuOptOut() throws {
         let (store, root) = try loadStore(globalJSON: """
         {
